Remove ap-start script
Summary:
The reason for the existence of ap-start.sh is that dhcpcd fires up
wpa_supplicant (which prevents hostapd from running properly), so the
script kills it before starting the AP stack.  The script already
suggested there should be a better mechanism, and this change implements
that.  Dhcpcd allows the configuration to disable specific hooks, one of
which is the wpa_supplicant hook.

Test Plan:
The problem that killing wpa_supplicant solves doesn't reproduce on my pi,
because the only interface managed by dhcpcd isn't wireless, so even though it
runs the wpa_supplicant hook, it doesn't run wpa_supplicant itself.  So I
confirmed that it did indeed run the hook if the `nohook` line wasn't present,
and that it didn't run when it was, and that the line did indeed get added on
package upgrade, as well as all the AP services starting up appropriately on
boot.

Reviewers: nils, stephen

Reviewed By: nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D367