Remove the non-20MHz channels from the lists of 20MHz channels
Summary:
The channel lists in networkd where we keep track of valid 20MHz
channels included channels numbers that were the centers for 40MHz
channels, and thus not valid 20MHz channels.  As part of that exercise,
it became apparent that there are two different ways to refer to a 40MHz
(or wider) channel -- either by its center or by its primary 20MHz
channel.  Our choice (which I'm not changing here) is to use the latter,
which matches the behavior in the MacOS UI, and simplifies creating
hostapd.conf.

Test Plan: None.

Reviewers: nils

Reviewed By: nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D368