Size of contents of the hostapd authTypes array is wrong
Summary:
The authTypes array is built by iterating over the rings, and appending
the authentication type for that ring.  With multiple rings using the
same authtype, this array may contain many duplicates.  The intent is
for the array to contain only a single entry per authentication type.

Test Plan: tested on Unielec board

Differential Revision: https://ph0.b10e.net/D370