Add ap-diag
Summary:
Add the `ap-diag` command, with one subcommand: `wifi`

The iwinfo module, formerly private to `ap.networkd`, is factored out into a new
package in `ap_common`, `wificaps`, and used by both `ap.networkd` and `ap-diag`.

Test Plan:
Run `ap-diag` on a system with valid and invalid wireless devices in it, and
verify that the output is correct, and looks decent.  Run the AP stack on the
same system and note that it comes up correctly, and properly logs the same
information as the command-line version.

Reviewers: stephen, nils, dp

Reviewed By: stephen, nils, dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D369