Reduce MAC footprint
Summary:
Instead of reserving a full 4-bit nybble on each interface for derived MAC
addresses, reserve the minimum needed.  On current platforms we only
have 2 SSIDs per interface (x-PSK and x-EAP), so we only need to reserve
a single bit.  Limiting our MAC footprint reduces the risk of two
different physical devices being reconfigured to use the same MAC
address.

Test Plan:
Verified that this works on USB devices on a Pi and on the miniPCI
devices on Demo Box A.

Reviewers: stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D381