Fix syntax validation for per-nic properties
Summary:
NICs are named differently on different platforms, but the syntax rules are
written assuming that NICs are all identified by the mac addresses.
This adds a new %nic% field type to the syntax rules, which will allow
either mac addresses or device names to be used.

Test Plan: running on both Pi and OpenWRT

Reviewers: stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D382