Add location to `ap-diag wifi` output
Summary:
We know some things about the physical locations of specific bus
addresses on the platforms we support, so we can emit that information
when it matches the information we have.

Fixes T130.

Test Plan: Run on Raspberry Pi and demo box, make sure output looks correct.

Reviewers: stephen, nils, dp

Reviewed By: stephen, nils

Subscribers: buildbot

Maniphest Tasks: T130

Differential Revision: https://ph0.b10e.net/D385