Add common/cfgtree
Summary:
This is primarily a refactoring changeset.  To support config trees being
managed and queried at multiple places in the cloud infrastructure, much of the
tree maintenance code has been pulled out of ap.configd and moved to a shared
common/cfgtree package.

We no longer attach an ops structure to every node in the tree, to support those
few nodes that need special handling when a property is being updated.  Instead,
on each set operation ap.configd will compare the path to those in the small
'special case' table.  This change allows for a cleaner separation between the
conceptual model of the tree as consumed by the daemons, and the implementation
of the tree in the shared module.

Each node has a hash value associated with it, representing a checksum of all
the nodes below it in the tree.

Test Plan:
This set of change has been exercised in isolation.  It has also been
used to implement a cloud-based cl.configd, which will appear in a subsequent
changeset.

All of the automated config tests now verify the correctness of the recalculated
root hash.

Reviewers: dp, stephen

Reviewed By: dp, stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D387