Add permanent MAC address to `ap-diag wifi` output
Summary:
Use the kernel's notion of what the "permanent" MAC address is.

Fixes T133.

Test Plan:
Compile on x86 and cross-compile to ARM.  Run on Raspberry Pi with a running
stack and see that the MAC address doesn't have the local bit set, and that the
bits we twiddle to allow for multiple SSIDs are untwiddled.  The addresses of
`wlanX` and `wlanX_1` match, and they match what was logged by `ap.networkd`.

Reviewers: dp, stephen, nils

Reviewed By: dp

Subscribers: buildbot

Maniphest Tasks: T133

Differential Revision: https://ph0.b10e.net/D389