ap.httpd: Reorganization of APIs
Summary:
- Try to align demo-api with anticipated cloud APIs; enables reuse of frontend when talking to cl.httpd, at least for now
- Extract authorization code into auth.go, /auth url namespace
- If user is logged in at application startup, skip login screen

To clarify for reviewers: this changes the URLs to better match what I think they will look like.  However it does not correct the 'store.js' implementation, which will need to become multi-appliance aware.  I've spent some time looking at the problem over the weekend, and that is going to be quite a bit more onerous-- it's not as simple as parameterizing the '/0/' part of the URL.

Test Plan: - Frontend testing

Reviewers: stephen, duvall

Reviewed By: stephen, duvall

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D393