cl.configd command queue
Summary:
Adds support for config SET, CREATE, and DELETE operations against the in-core
config tree.  Changes made to this cached tree are not persisted to disk.

This adds a command queue to cl.configd.  SET, CREATE, and DELETE are added to a
command queue, and the submitting rpc returns with an INPROGRESS status.  The
submitter can poll for command completion using the Status() rpc.  The command
queue currently lives completely in-core.  Eventually this queue will be
implemented via a database, allowing it to survive daemon reboots, migrate
between cl.configd instances, and be used as a changelog.

When we load an appliance's config from a file, we spin up a go routine that
emulates the behavior of a remote appliance.  This thread will periodically pull
submitted commands from the command queue, execute them against the in-core
tree, and pushes the results to a completion queue.  There are delays between
pulling events from the queue and pushing results, meant to mimic the behaviour
of the commands being processed by a remote appliance.

Test Plan: manual testing on tty0

Reviewers: dp, stephen

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D395