Add cfgapi
Summary:
This changeset adds a new common/cfgapi module that presents a Go interface for
interacting with a config daemon.  This allows the same code to be used on the
appliance to interact with ap.configd and in the cloud to interact with
cl.configd.  Much of the code in cfgapi has simply migrated from ap_common/apcfg
to the new common module.

Each platform must provide a module that exports the ConfigExec interface, which
encapsulates the interdaemon communication mechanism.  On the appliance, this is
implemented by ap_common/apcfg, which communicates over the 0MQ bus.  In the
cloud, it is implemented by cl_common/clcfg, which communicates using gRPC.

With the new API, the implementations of ap-configctl and cl-configctl
have been almost entirely merged.  These two utilities now just set up their
cfgapi links and jump into a new shared common/configctl module, which does all
of the command and response parsing.

Test Plan:
running full system on appliance and cloud.  Verified that
cl-configctl works in the cloud, ap-configctl works everywhere.

Differential Revision: https://ph0.b10e.net/D406