Make sure briefpg uses a consistent DB user
Summary:
If `$PGUSER` is set in the environment, then the tests using `briefpg`
fail.  `initdb` creates a database superuser matching the unix username,
regardless of whether `$PGUSER` is set, but `psql` connects to the
database using `$PGUSER`, if set, and the unix user if not.

So if the variable isn't set, it all works as you'd expect, simply
creating and using the user `duvall` or `jenkins` or whatever.  If it is
set, and set to something other than the unix user, then the database
connection fails with an error:

    FATAL:  role "postgres" does not exist

which, depending on which piece of which test it occurs in, may or may
not be displayed as part of the test failure output.

This change forces the user to be consistently `postgres`.

It also removes the URL escaping for the socket path in the `host`
parameter in the DB URI.  That's only necessary when putting a socket
path in the standard URI host position.

Test Plan:
Run `make test` with `$PGUSER` unset, set to `duvall`, set to `postgres`,
set to `junk`, and set to ``, and never see the tests fail (also with
`GOCACHE=off` set to make sure the tests get re-run regardless of past
success).

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D407