Change appliance_pubkey.expiration to have a time zone
Summary:
The `expiration` column in the `appliance_pubkey` database is a timestamp, but
by default that data type doesn't have an associated timezone, so isn't
especially useful.  We can change the type of the column in the existing
database with a simple SQL statement:

    ALTER TABLE appliance_pubkey ALTER COLUMN expiration TYPE timestamp with time zone;

since the column is currently unused, and has no values that need to be
corrected.

Test Plan: Test suite.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D408