Allow multiple config stores in cloud
Summary:
As part of the review of 1a8beb6, Nils asked for the ability to override
the database config store with local files, to be used during
development.

Test Plan:
Manual: run `cl-configctl` to retrieve data for appliances that are in the file
store but not in the database, and vice versa.  Also make sure that passwords
are properly censored for a variety of database connection URLs.

Reviewers: dp, nils

Reviewed By: dp, nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D410