Cloud and appliance use hashes to enforce config synchronization
Summary:

ap.configd issues config updates in the order they were applied.
Previously the updates from a multi-step operation were issued in a
non-deterministic order.  This worked for our on-appliance purposes, but
is insufficient when the cloud is using the updates to keep its cached
tree in sync with the on-appliance tree. As a side effect of this
change, we can now implement notifications without needing the 'notify
callbacks' in the cfgapi interface.

Hashes are recomputed for each property change - not just when all
operations in a multi-step operation complete.  This allows us to issue
each update individually with a correct hash.  This allows the cloud to
verify the results of each update, not just when a batch of updates are
all completed.

If cl.configd detects a hash mismatch after applying a property update,
it will request a fresh tree from the appliance.

Test Plan: currently running with local nilsap and remote svc0

Differential Revision: https://ph0.b10e.net/D421