Make svc rpc address a config property rather than a hardcoded value
Summary:

Instead of relying on a hardcoded URL to find the rpc daemon, the
address now comes from the config file.  We stil have a hardcoded
default to handle the chicken-and-egg problem of recovering from a lost
config.  We also added a mechanism for cl.rpcd to instruct ap.rpcd to
move to a new cl.rpcd address, allowing us to offload work if cl.rpcd
gets overloaded with too many appliances using that hardcoded default.

Unrelated bug fix: when handling a Net.Entity event in configd, we need
to calculate the hash values of each updated property prior to sending
the associated notification to ap.brokerd. This was casusing hash
mismatches to be deteced in the cloud.

Test plan: Running on nilsap and svc0

Differential Revision: https://ph0.b10e.net/D428