Fix upgrade-related hash bugs
Summary:
When adding a new node in cfgtree.Move(), we need to compute the hashes
from the new node to the root before adding more children to that node.
After a subtree has been moved, we need to recompute the hashes from
that subtree's new location to the root.

We need to do a ChangesetInit() / ChangesetCommit() around a config
tree upgrade to ensure that any rollback state that gets created gets
cleaned up.

Add a version upgrade to the ap.configd test suite - specifically
including an upgrade that involves a subtree move.

Test Plan: added new tests to the test suite

Reviewers: dp

Reviewed By: dp

Subscribers: dp, buildbot

Differential Revision: https://ph0.b10e.net/D427