database-backed command queue
Summary:
The command queue as presently implemented is entirely in-core, which means that
if `cl.configd` shuts down, the queue disappears.  The queue needs to have some
sort of persistent backing, and given the nature of what we're trying to do and
the tools at our disposal, a database-backed queue seems to make the most sense.

Test Plan:
Manual testing with `cl-configctl` and the appliance emulator embedded in
`cl.configd`.

Reviewers: nils, dp

Reviewed By: nils, dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D423