Fix panic in apstate.store() when cachedTree has not yet been initialized
Summary:
This resolves the panic we were seeing in the November 2 rollout.  A series of
updates which race ahead of a refresh are "successfully" applied (they are
discarded by backend.go:update())-- as a result of the success, the code thinks
that the tree needs to be saved; so it calls perAPState.store() which tries to
store a nil tree.

Test Plan:
Ran on svc1 all weekend; some logs were lost due to verbose logging I added,
but I do not think it has not crashed since this fix was applied.

Reviewers: duvall, nils

Reviewed By: duvall, nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D437