Introduce cl-reg, the registry tool
Summary:
It has a two-deep subcommand structure, with the first level accepting
`app` (for appliance), and the second level accepting `add` and `list`.
The expectation is that `reg` would be a future first-level subcommand,
allowing you to manage entire registries.

The intent is to replace `build/gcp-appliance-reg/new-appliance.sh` with
a Go program, and to separate out the core logic so that it can be
reused by daemons wishing to provide another front-end.

This also introduces the ability to create arbitrary transaction
boundaries in the appliance DB routines.

Fixes T138.

Test Plan: Manual, against a test database.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Maniphest Tasks: T138

Differential Revision: https://ph0.b10e.net/D429