Stackdriver logging
Summary:
With the `-log-type` flag set to `stackdriver`, and an appropriate credentials
file provided, the cloud daemons will log directly to Google's Stackdriver
Logging, where logs can be queried in a structured way, simple metrics can be
created, and the logs then further dumped to GCS, BigQuery, or PubSub.

The logs continue to be sent to stderr in a zap JSON encoding.

There are also fixes for cosmetic problems in a couple of messages, and a fix
for a context issue with the appliance emulator.

Test Plan:
Manual testing on personal-dd0, with an emulated appliance, and inspecting the
generated logs in the cloud console, as well as with `gcloud logging read`.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Maniphest Tasks: T147

Differential Revision: https://ph0.b10e.net/D451