Add profiling interfaces
Summary:
This adds golang profiling targets to most of the daemons on both cloud
and appliance.  It also adds a webserver to the few daemons that did not
already include one for Prometheus stats collection.  Because the same
port is being used for prometheus and profiling, they have been renamed
"DIAG" ports.

This removes the hand-rolled memory and CPU profiling that were
supported by watchd and identifierd.

For usage instructions, see:
       https://blog.golang.org/profiling-go-programs

Test Plan: manual on both a local pi and tty0

Differential Revision: https://ph0.b10e.net/D468