Limit rings subject to anti-phishing DNS
Summary:
We only have access to hostnames, not full URLs, so our DNS-based
anti-phishing mechanism is too heavy-handed to be used for all systems.
With this change, only those systems in the Devices, Unenrolled, and
Quarantine rings will be subject to our anti-phishing limitations.
Systems in the remaining rings are expected to use some local
antiphishing tools that can operate with more precision.

Test Plan: manual testing

Reviewers: stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D492