Add ssh service tunnel support
Summary:
Adds a cl-service utility that sets config tree properties describing an
ssh tunnel from the appliance to the cloud.  cl-service then spawns an sshd
daemon which will serve as the cloud endpoint for the tunnel.

ap.rpcd detects the new config tree properties and responds by spawning an
ssh daemon that will serve as the appliance endpoint for the tunnel.  It
then establishes an ssh connection to the cloud sshd daemon, and opens the
tunnel over that connection.  As new ssh clients connect to the tunnel port
in the cloud, ap.rpcd plumbs those new connections through to the sshd
daemon running on the appliance.

Test Plan: manual testing between nilsap and z-tun0.

Differential Revision: https://ph0.b10e.net/D515