Fetch user phone number from Google People API (when available)
Summary:
Populate the user phone number from the google People API when it is
available there, at account creation.  In the future we can attempt to capture
the phone number at every login when it is missing, but this is not supported
yet.

Test Plan:
Added a phone number to the test domain account, and checked that it
flowed correctly through new account creation.

Reviewers: duvall, stephen

Reviewed By: duvall

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D528