Extricate cl.httpd /api/account endpoints from /api/site endpoints
Summary:
This pulls the account-related endpoints apart from the site-related
endpoints in a much more complete way than we had before.  This is
almost entirely reorganization, but it makes the code easier to deal
with going forward.

Test Plan:
Ran fleet of cloud daemons and checked that functionality remains
the same in the browser app.  Also spot-checked APIs with a logged-out
user to make sure middleware is installed properly.

Reviewers: stephen, duvall

Reviewed By: duvall

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D522