OpenWrt updates for Product Software
Summary:
This changeset moves the sysroot and toolchain to the current OpenWrt
build.  It also introduces preremove and postinstall configuration
changes to modify root's interactive `PATH` and to make rsyslog behave
similarly to Debian with respect to configuration file locations.
Finally, it adds knowledge about chronyd differences to the platform
definition.

Test Plan: Tested install and remove on MT7623.

Reviewers: dp, duvall, nils

Differential Revision: https://ph0.b10e.net/D520