T180 Begin IPv4 range management via site_index
Summary:
Rather than having subnets configured for each ring, the subnets are
now calculated based on a starting address and a site index.

Add a new @/site_index property to the config tree, which is used to
distinguish between sites within the same organization.

Add a new @/network/base_address property, which is used as a starting
point when calculating the IP ranges assigned to each ring.  The choice
of a base_address allows us to determine which of the different private
address ranges to use.

There is also some minor refactoring, as shared network code is moved
from ap_common/network to common/network.  As all of the remaining AP
code is related to dhcp, the module has been renamed ap_common/dhcp.

Test Plan: tested with both new and upgrade configs on nilsap.

Differential Revision: https://ph0.b10e.net/D529