Appliance certificate lifecycle management
Summary:
This introduces `cl-cert`, which is the cloud program responsible for managing
TLS certificates for sites.  This includes requesting them from Let's Encrypt,
storing them in the database, and renewing them when appropriate.  It is run
periodically via systemd.

There is a new RPC connection between `ap.rpcd` and `cl.rpcd`, which is used to
request and download a certificate.

Certs are requested before they're needed by site, and stored in the database in
a pool of unclaimed certs.  When a site's gateway appliance pings the cloud for
the first time, it will request a certificate, and one will be claimed from the
pool for that site (whose domain will then be changed to match that of the
certificate).

When a certificate is claimed, a property is added to the config tree:
`@/certs/<hash>/state`, where `<hash>` is the hash of the certificate, the value
of the property is `available`, and the expiration is set to the end of the
certificate's validity period.  The appliance listens for changes on that path
(as happens when `cl-cert` renews a certificate) and will download and install a
certificate when it becomes available, changing the property to `installed`, and
notifying any daemons which care about it.

Closes T171

Test Plan:
There is automated testing at both the SQL interface level (as we have for all
the `DataStore` interfaces) and for most of the higher level methods in
`cl-cert` itself.

Reviewers: dp, stephen, nils

Reviewed By: dp, stephen, nils

Subscribers: buildbot

Maniphest Tasks: T171

Differential Revision: https://ph0.b10e.net/D526