Cache download of oui.txt
Summary:
Builds have been failing all day because fetching `oui.txt` from
http://standards-oui.ieee.org/oui.txt has been failing with a connect timeout.
It's not consistent, but enough of a pain that we need to get past it somehow.
Caching the file seems like a reasonable enough first response.

Test Plan:
Run the build with and without `$DOWNLOAD_CACHEDIR` set, with it unpopulated and
pre-populated, and see what happens when the curl command succeeds and fails.

Reviewers: dp, stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D548