Add GetVirtualAPs to cfgapi
Summary:
The code used to parse the @/network/vap portion of the config tree is
generally useful, so it has been moved out of ap.networkd into
common/cfgapi.

Test Plan:
running on OpenWrt board with all VAPs in use, with no rings
assigned to the guest VAP, and with EAP parameters missing.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D551