Create private key PEM files OpenSSL can understand
Summary:
When `userauthd` loaded the cloud-sourced private key, it complained bitterly
about the file format.  It turns out that the PEM header was "BEGIN PRIVATE KEY"
instead of "BEGIN RSA PRIVATE KEY".  This commit changes the code to emit the
latter.

Test Plan:
Hand-edit the private key on dogfood to add "RSA" and verify that userauthd can
now load the file.  Also see that `openssl pkey -noout -text < privkey.pem`
doesn't complain.  With the changes in place, on a cloud-based abbreviated
appliance, see that a newly downloaded key gets written to disk with the correct
header and that the `openssl` command doesn't choke.

Reviewers: nils, dp, stephen

Reviewed By: nils, dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D558