Don't increment max_claimed when nothing actually gets claimed
Summary:
When an appliance requests a certificate, it gets assigned a siteid,
whether there's a certificate for it or not.  If there isn't, it will
keep trying, and although it doesn't get assigned a new siteid each
time, the `max_claimed` value in the `siteid_sequences` table gets
updated.  This is because the trigger that updated that table ran
regardless of whether the insertion into `site_domains` succeeded.
Given that we need both a conflict check on insertion and to change the
value that we insert into the table (i.e., add the siteid we generate),
a trigger won't work for us.  Instead, use a function that operates in a
single transaction.  This simplifies the `RegisterDomain()` function,
too, as well as fixing a bug in it where a re-registration always
returns the domain for siteid 0.

This changeset includes a change to the output of `cl-cert list` to emit
the expiration in the local time rather than the stored time (UTC).

Test Plan: Added automated tests for the issues getting fixed.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D563