Support config restoration from the cloud
Summary:
When ap.rpcd connects to the cloud for the first time, it should check
whether the appliance already has a config tree backed up in the cloud.
If so, the tree should be downloaded and used rather than constructing a
new ap_props.json using the standard default settings.

[ap|cl]-configctl has a new 'replace' command, which will replace any
existing config tree with one provided by the user.

[ap|cl]-configctl has a new 'export' command, which will dump a json
representation of the config tree in a format that can be used by the
'replace' command.

Test Plan: manual testing on svc0 and local nilsap appliance

Differential Revision: https://ph0.b10e.net/D565