Add command queue functionality to cl-reg
Summary:
The command queue has lots of useful and interesting data, and is rich enough
that invoking `psql` and typing a bunch of SQL to extract it is annoying and
difficult to remember.  This gives us most of what we need, and can be extended
with more functionality as required.

Closes T207

Test Plan: Ran against the queue on `svc1`.

Reviewers: dp, nils

Reviewed By: dp

Subscribers: buildbot

Maniphest Tasks: T207

Differential Revision: https://ph0.b10e.net/D566