T230 Alpha behavior for cloud accounts / roles and low-privilege users
Summary:
This commit introduces cloud account roles.  Roles are expressed as a tuple of
{account, org, role} to be read as "With respect to organization {org} Account
{account} has role {role}." Roles can in the future be extended to
appliance-local activities, however this is not implemented as of this commit.

For now, roles are managed via cl-reg, except that new users are automatically
given the "user" role; we will manually assign roles to the existing user population.

cl.httpd has been modified to enforce the roles for the site APIs.

The frontend has been modified to adapt its behavior depending on the
roles the user has.  Since the frontend currently lacks the concept
of an organization, this is presently expressed per site.

Test Plan:
Manual testing:
- Assign and remove roles from users, observe behavior
- Web frontend testing
- Inspected results of REST API to be sure that the right behavior is seen

Reviewers: duvall, stephen

Reviewed By: duvall, stephen

Subscribers: buildbot

Maniphest Tasks: T230

Differential Revision: https://ph0.b10e.net/D564