Put site_uuid and appliance_uuid into cloud logs
Summary:
Make `endpointLogger()` in `cl.rpcd` common, and have both `cl.rpcd` and
`cl.configd` use it wherever appropriate, which means extracting the fields out
of the gRPC headers in the context in most functions in the stack, and adding
them to the sugared logger.  We assume in many places this is very cheap.

In a couple of places, context arguments needed to be added, previously `nil`
contexts changed to use those.

`cl.eventd` is not a gRPC service, so instead it pulls the fields out of the
message from the bus.

Fixes T184

Test Plan:
Manually check that the logs for `cl.configd`, `cl.eventd`, and `cl.rpcd` all
have `site_uuid` and `appliance_uuid` on every entry that's associated with an
appliance or a site.

Reviewers: dp, nils

Reviewed By: dp, nils

Subscribers: buildbot

Maniphest Tasks: T184

Differential Revision: https://ph0.b10e.net/D569