Rename compound-noun properties to use snake_case
Summary:
On config upgrade, rename a handful of properties that use dashes or camelCase
to join words to instead use underscore.

Fixes T209

Test Plan:
With most of the properties provisioned in the tree, run the new bits to force
the upgrade, and see that the property upgrade show up in the log, and that the
new config store looks as expected.

Reviewers: dp, stephen, nils

Reviewed By: stephen, nils

Subscribers: buildbot

Maniphest Tasks: T209

Differential Revision: https://ph0.b10e.net/D571