Add kernel log monitoring and actions
Summary:
Add a REBOOT command to mcp, which reboots the whole platform.

ap.logd now watches the kernel log, and will take action when it sees
specific messages.  Currently it watches for a special test message,
which causes our daemons to restart, and an error message from the mtk
switch (T256), which causes us to reboot the platform.

Test Plan:
Running on OpenWRT board.  Tested both the special message and
artificially injected the switch error.

Differential Revision: https://ph0.b10e.net/D577