Compute Go dependencies
Summary:
This removes from the Makefile the hand-crafted dependencies between programs and individual `.go` files, and replaces them with an auto-generated makefile (`deps.mk`). The idea is that for each binary, it discovers what Go packages it depends on, and for each package, it discovers what `.go` files constitute it, and emits the dependency graph in `make` form. It doesn't allow for finer-grained dependencies (such as only some programs relying on the `certs.go` portion of the `appliancedb` package instead of all programs which import that package), but that is conservative and safe.

Fixes T75

Test Plan: Manual

Reviewers: dp, stephen

Reviewed By: dp, stephen

Subscribers: buildbot

Maniphest Tasks: T75

Differential Revision: https://ph0.b10e.net/D552