T175 Move important mutable appliance files to /data
Summary:
This changeset reorganizes our mutable files such that they can be moved
on a per-platform basis.  For the "mediatek" platform, this capability
means that we place the mutable state and appliance logs on `/data`,
which is a larger partition that is shared by both bootable sides.

The logic is that there are four expanding tokens: `__APROOT__`,
`__APPACKAGE__`, `__APDATA__`, and `__APSECRET__`.  These are described in the
new version of `README.md`.  Once a component has identified its
platform, it can use `.ExpandDirPath()` to compute the final path.

Note that this changeset unifies the meaning of `APROOT` across all
components.

Test Plan:
Built, and packages built for both distros.  OpenWrt packages installed
and tested.  Debian packages installed and tested.

Reviewers: duvall, dp, nils

Differential Revision: https://ph0.b10e.net/D582