Rediscovered dhcp lease results in "out of leases"
Summary:

Prior to commit e0ab8e6f, leaseAssign() only returned a pointer to a lease.
A nil pointer was interpreted as an error. With that commit, we now return
an explicit error. That error should only be set if the pointer being
returned is nil. Currently it can also be set if we rediscover a stale
lease in the table and attempt to reassign it to its original owner.

Test Plan: Manual testing on nils-plateau.

Differential Revision: https://ph0.b10e.net/D587