Stop spurious "declining to move" warnings
Summary:
When a device is assigned to a non-default ring on a VAP, on each
connection attempt configd will warn that it is "declining to move" the
device to that default ring.  This message should only be shown when
configd is declining to move a device's ring assignment between VAPs.

Add Devices->Standard as an acceptable ring transition, so there are
automatic transitions from both PSK networks to EAP.

Test Plan: manual

Differential Revision: https://ph0.b10e.net/D590