T299 ap.serviced relaying stops when the bridges are rebuilt
Summary:
When networkd rebuilds the bridges, the relay goroutines need to be stopped
and restarted so the new interfaces can be added to the UDP multicast
groupd.

Remove frequent, but uninformative, message from urlfetch.

Increase default lease times for the devices and guest rings.

hostapd has to watch for CTRL-EVENT-EAP-RETRANSMIT2 messages as well as
CTRL-EVENT-EAP-RETRANSMIT.

Test Plan:
manual with nils-plateau, a macbook, and both an HP printer and
spotify receiver.

Reviewers: stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D592