remove appliance-side tensorflow
Summary:
Because `ap.identifierd` combines collection and inference, and there is
no TensorFlow for our ARM processor, we have not been collecting and
uploading device data from MT7623-based gateways.  This changeset
excises the TensorFlow dependency so that collection can resume.

Test Plan:
Installing on various MT7623 gateways.  Waiting to see if new uploads
are forthcoming, or if `ap.identifierd` and `ap.rpcd` are out of sync
because of `/data`.

Reviewers: dp, duvall

Differential Revision: https://ph0.b10e.net/D598