T298 firewall messages should be dropped from /var/log/messages
Summary:
On nils-plateau, /var/log/messages was more than 99% full of DROPPED
messages generated by iptables.  These messages are consumed by watchd,
which uses them to monitor firewall activity.  This change will update
the rsyslog config so these messages are available to watchd, but are
not delivered to networkd or persisted in /var/log/messages.  This
reduces the CPU load on networkd and wear on our flash storage.

While watchd is portscanning a device, we are blocking all returning UDP
traffic and some returning TCP traffic.  These blocks result in multiple
DROPPED messages, which may make it more difficult to detect legitimate
firewall violations.  In addition, the missing return traffic may result
in nmap drawing incorrect conclusions about the target device.  This
change will remove the firewall barrier between a client and the
appliance for the duration of a portscan operation.

Test Plan: manual

Differential Revision: https://ph0.b10e.net/D599