Track RETRANSMIT event per-client
Summary:
With our current wifi hardware, hostapd can get into a state where it
claims to be RETRANSMITting EAP packets to the radius server, but no
such operation is actually taking place.  To break it out of this loop,
we restart hostapd.  If two clients are attempting to connect at the
same time we can hit the restart threshold prematurely, potentially
causing an unnecessary network interruption.  To prevent this, we now
track RETRANSMIT events per-client, and only perform the reset once
one client has seen enough events to warrant the reset.

Unrelated: change the log level of DROPPED packet messages from the
default 'warn' to 'info'.  This prevents the messages from being
displayed on the console.

Test Plan: manual on nils-wrt

Differential Revision: https://ph0.b10e.net/D602