brvlan0 should be populated by logical devices, not physical
Summary:
Unenrolled clients are assigned to brvlan0.  The unenrolled ring may be
provided by wlanX, wlanX_1, or wlanX_2, but we are currently always
putting wlanX on the vlan0 bridge.  The fix is to use the name of the
logical devices when populating the ring rather than the physical.

Three small bug fixes:

  - When deleting a config subtree, we were repeatedly sending
    notifications about the root node being deleted - not each of the
    child nodes.

  - Don't attempt to attach a nic to brvlan0 until the nic has been
    created.  This was harmless, but caused unnecessary error messages
    to be printed.

  - Don't attempt to update DNS records for non-existent clients.

Test Plan:
manual with clients connecting to different VAPs with and
without being explicitly assigned to rings.

Differential Revision: https://ph0.b10e.net/D613