Overwrite the existing chrony init script with our own
Summary:
Getting the OpenWRT build to override the built-in `chrony` init and hotplug
scripts was too crazy, so for now just overwrite the former with our own copy
and remove the latter.

Test Plan:
Give `.ipk` package to @stephen, update package on test appliance, verify that
`chrony` got restarted (previous fix) and config file was updated.

Reviewers: stephen

Reviewed By: stephen

Subscribers: buildbot, stephen

Differential Revision: https://ph0.b10e.net/D614