Startup robustness on OpenWRT.  Dynamic transition from gateway to satellite mode.
Summary:
On the OpenWRT platform, our software stack can come fully online before
we have a DHCP address.  Because we rely on data in the DHCP response to
determine whether to act as a satellite, we don't know which mode our
software should start in.  Rather than blocking until we have that info,
which could hinder debugging a connectivity issue, by default we will
start in gateway mode.  We will transition to satellite mode if a
subsequent DHCP transaction instructs us to.

Some daemons (specifically watchd and networkd) may take more than than
the default 100ms to shut down cleanly.  If ap.mcp kills the during
shutdown, we may be left with a random subset of bridges and interfaces
still active.  This can confuse the OpenWRT dhcp daemon, and it
increases the likelihood of an error if/when the daemons restart.   It
is now possible to specify a per-daemon timeout, giving those daemons
more time.

If two wireless devices have overlapping MAC address ranges, issue a
warning and leave the second offline.

Test Plan:
multiple restarts/reboots of an OpenWRT board as a satellite
node with an RPI gateway.

Differential Revision: https://ph0.b10e.net/D626